// apps/frontend/core-dapp/src/types/market.types.ts
export type Range = [number, number]

export type SupportedMarketID = 'arbitrum' | 'zksync' | 'starknet'

export type MarketType = 'preICO' | 'preIPO'

export type MarketVersion = 'v1.0' | 'v1.1'

export type Market = {
  iconName: IconName
  type: MarketType
  name: string
  urlId: SupportedMarketID
  companyName: string
  totalSupply?: number
  address: SupportedMarkets
  valuationMultiplier?: number
  hideExpiry?: boolean
  long: {
    tokenAddress: SupportedMarketTokens
    poolAddress: SupportedMarketPools
  }
  short: {
    tokenAddress: SupportedMarketTokens
    poolAddress: SupportedMarketPools
  }
  static: {
    valuationRange: Range
  }
  version: MarketVersion
  settlementDocsLink?: string
}